feat: move onsite/exhibition/status paths to exhibition
