+++ src/apis/exhibitions/index.ts
@@ -1,4 +1,4 @@
-import { AuthToken } from "../../../@types";
+import { AuthToken } from "../@types";
 import { AllStatus } from "./@types";
 
 export interface Methods {
